New tracking system
You can now lock with a double-click on anything, food included. This
also fixes a few bugs and subtly changes the way clicking and
double-clicking works.

+++ petridish/src/main/java/org/nusco/narjillos/views/RoundObjectView.java
@@ -10,9 +10,9 @@ import org.nusco.narjillos.utilities.Viewport;
 public abstract class RoundObjectView extends ThingView {
 
 	private final Circle shape;
-	private final int radius;
+	private final double radius;
 
-	public RoundObjectView(Thing thing, int radius) {
+	public RoundObjectView(Thing thing, double radius) {
 		super(thing);
 		this.shape = new Circle(radius);
 		this.radius = radius;
